requirements.txt
google-generativeai
typing-extensions
reportlab